separating sublime package from sublime plugin
