+++ lib/sublime/package.go
@@ -1,0 +1,10 @@
+package sublime
+
+type pkg struct {
+	text.HasSettings
+	keys.HasKeyBindings
+	platformSet *text.HasSettings
+	defaultSet  *text.HasSettings
+	defaultKB   *keys.HasKeyBindings
+	plugins     []plugin
+}
