--- lib/sublime/sublime.go
@@ -1,7 +1,0 @@
-package sublime
-
-type (
-	build struct{}
-
-	snippet struct{}
-)
